+++ Resources/public/js/util/price-calculation-util.js
@@ -199,23 +199,32 @@ define([], function() {
                 var itemTax = 0;
                 var ratio = 0;
                 var tax;
+                var numberItems = retrieveNumberOfNonRecurringItems(items);
 
                 // Calculate taxes for delivery costs in relation to the item prices.
                 for (var i in items) {
+                    item = items[i];
+
+                    if (item.isRecurringPrice) {
+                        continue;
+                    }
+
                     if (areValidCalculationParams(
                             sandbox,
-                            items[i].price,
-                            items[i].tax,
-                            items[i].discount,
-                            items[i].quantity)
+                            item.price,
+                            item.tax,
+                            item.discount,
+                            item.quantity)
                     ) {
-                        item = items[i];
                         itemNetPrice = calculateNetPrice(sandbox, item);
                         itemTax = item.tax.toString();
 
                         ratio = 0;
                         if (result.netPrice != 0) {
                             ratio = itemNetPrice / result.netPrice;
+                        } else if (numberItems > 0) {
+                            // Handle total net price of 0. Each item has the same ratio.
+                            ratio = 1 / numberItems;
                         }
 
                         tax = ratio * deliveryCost * item.tax / 100;
@@ -232,14 +241,32 @@ define([], function() {
         },
 
         /**
+         * Returns number of items that do not have recurring prices.
+         */
+        retrieveNumberOfNonRecurringItems = function(items) {
+            var numberItems = 0;
+
+            for (var i in items) {
+                var item = items[i];
+                if (item.isRecurringPrice) {
+                    continue;
+                }
+                numberItems++;
+            }
+
+            return numberItems;
+        },
+
+        /**
          * Processes elements for getTotalPricesAndTaxes.
          * Returns null if one value is invalid.
          *
          * @param {Object} sandbox
          * @param {Array} items
          * @param {Number} deliveryCost
+         * @param {Bool} shouldCalculateRecurringPrices
          */
-        processPriceCalculationItem = function(sandbox, items, deliveryCost) {
+        processPriceCalculationItem = function(sandbox, items, deliveryCost, shouldCalculateRecurringPrices) {
             var tax = 0, i, item, itemNetPrice,
                 result = {
                     taxes: {},
@@ -247,20 +274,28 @@ define([], function() {
                     grossPrice: 0
                 };
 
+            shouldCalculateRecurringPrices = !!shouldCalculateRecurringPrices;
+
             if (typeof deliveryCost === 'undefined') {
                 deliveryCost = 0;
             }
 
             try {
                 for (i in items) {
+                    item = items[i];
+
+                    // Check if price is recurring or not.
+                    if (item.isRecurringPrice !== shouldCalculateRecurringPrices) {
+                        continue;
+                    }
+
                     if (areValidCalculationParams(
                             sandbox,
-                            items[i].price,
-                            items[i].tax,
-                            items[i].discount,
-                            items[i].quantity)
+                            item.price,
+                            item.tax,
+                            item.discount,
+                            item.quantity)
                     ) {
-                        item = items[i];
                         itemNetPrice = calculateNetPrice(sandbox, item);
 
                         tax = itemNetPrice * getTaxRate(sandbox, item.tax);
@@ -440,6 +475,7 @@ define([], function() {
          * @param {Object} sandbox
          * @param {Object} items
          * @param {Float} deliveryCost
+         * @param {Bool} shouldCalculateRecurringPrices
          * [
          *  {
          *   price: 100,
@@ -450,9 +486,9 @@ define([], function() {
          *
          * @return {Object}
          */
-        getTotalPricesAndTaxes: function(sandbox, items, deliveryCost) {
+        getTotalPricesAndTaxes: function(sandbox, items, deliveryCost, shouldCalculateRecurringPrices) {
             if (!!items) {
-                return processPriceCalculationItem.call(this, sandbox, items, deliveryCost);
+                return processPriceCalculationItem.call(this, sandbox, items, deliveryCost, shouldCalculateRecurringPrices);
             }
             return null;
         },